DRY up tests with custom logger
